Changes to using decimal for coordinates from integer

+++ db/migrate/20190223003501_create_cities.rb
@@ -2,8 +2,8 @@ class CreateCities < ActiveRecord::Migration[5.1]
   def change
     create_table :cities do |t|
       t.string :name
-      t.integer :latitude
-      t.integer :longitude
+      t.decimal :latitude
+      t.decimal :longitude
       t.string :image
 
       t.timestamps
